lazy & dedupe
