+++ package.json
@@ -6,7 +6,7 @@
     "format": "prettier \"./projects/core/**\" --write",
     "test": "vitest --ui --coverage",
     "release": "rimraf dist && npm run build && npm run patch-versions",
-    "build": "tsc -p projects/core/tsconfig.build.json && tsc -p projects/core/tsconfig.build.esm.json && tsx misc/build-packagejson-core.ts && copyfiles README.md dist/remult && copyfiles LICENSE dist/remult && tsx misc/extract-public-api.ts",
+    "build": "tsc -p projects/core/tsconfig.build.json && tsc -p projects/core/tsconfig.build.esm.json && tsx misc/build-packagejson-core.ts && copyfiles README.md dist/remult && copyfiles LICENSE dist/remult && tsx misc/dedupe-admin-html.ts && tsx misc/extract-public-api.ts",
     "build-tests": "tsc -p tsconfig.tests.json && tsc -p projects/tests/strict/tsconfig.check.json",
     "build-watch": "concurrently \"tsc -w -p projects/core/tsconfig.build.esm.json\" \"tsc -w -p projects/core/tsconfig.build.json\"",
     "all-servers": "concurrently -k -n \"koa,fastify,express,mw,deno-fresh,mwc,next,svelteKit,next-app\" -c \"bgBlue.bold,bgGreen.bold,bgCyan.bold,bgYellow.bold\" \"tsx --tsconfig tsconfig.server.json projects/test-servers/node-servers/koaServer.ts\" \"npm run fastify-server\" \"tsx --tsconfig tsconfig.server.json projects/test-servers/node-servers/expressServer.ts\" \"tsx --tsconfig tsconfig.server.json projects/test-servers/node-servers/middlewareServer.ts\" \"cd projects/test-servers/deno-fresh && deno task start\" \"tsx --tsconfig tsconfig.server.json projects/test-servers/node-servers/middlewareCore.ts\" \"cd projects/test-servers/next-server && npm run dev\" \"cd projects/test-servers/sveltekit-server && npm run dev\" \"cd projects/test-servers/next-app-server && npm run dev\" ",
